first try to merge AllChecksRunner and ChecksRunner
